adds signin test

+++ controllers/contacts-controller.js
@@ -1,4 +1,5 @@
 const fs = require('fs/promises');
+const gravatar = require('gravatar');
 
 const { Contact } = require('../models/Contacts');
 const { HttpError, cloudinary } = require('../helpers/index');
@@ -9,6 +10,7 @@ const { ctrlWrapper } = require('../decorators/index');
 const getAll = async (req, res) => {
   const { _id: owner } = req.user;
   const { page = 1, limit = 10 } = req.query;
+
   // skip - сколько объектов нужно пропустить c начала коллекции
   // limit- сколько объектов нужно вернуть с коллекции
   const skip = (page - 1) * limit;
@@ -16,6 +18,7 @@ const getAll = async (req, res) => {
     skip,
     limit,
   }).populate('owner', 'email subscription');
+
   res.json(result);
 };
 
@@ -23,25 +26,40 @@ const getById = async (req, res) => {
   // req.params - это объект который содержит все динамические части запроса
   const { contactId } = req.params;
   const { _id: owner } = req.user;
+
   // const result = await Contact.findById(contactId);
   const result = await Contact.findOne({ _id: contactId, owner });
   if (!result) {
     throw HttpError(404, `Contact with id=${contactId} not found`);
   }
+
   res.json(result);
 };
 
 const addContact = async (req, res) => {
   const { _id: owner } = req.user;
-  const { url: fotoURL } = await cloudinary.uploader.upload(req.file.path, {
-    folder: 'contacts-project/users',
-  });
-  await fs.unlink(req.file.path);
+  let fotoURL;
+
+  if (req.file) {
+    // const { url: fotoURL } = await cloudinary.uploader.upload(req.file.path, {
+    //   folder: 'contacts-project/users',
+    // });
+    const cloudinaryResponse = await cloudinary.uploader.upload(req.file.path, {
+      folder: 'contacts-project/users',
+    });
+    fotoURL = cloudinaryResponse.secure_url;
+    await fs.unlink(req.file.path);
+  } else {
+    const { email } = req.body;
+    fotoURL = gravatar.url(email, { s: '200', d: 'identicon' });
+  }
+
   const result = await Contact.create({
     ...req.body,
     fotoURL,
     owner,
   });
+
   res.status(201).json(result);
 };
 
@@ -49,6 +67,7 @@ const updateById = async (req, res) => {
   const { contactId } = req.params;
   const { _id: owner } = req.user;
   // const result = await Contact.findByIdAndUpdate(contactId, req.body);
+
   const result = await Contact.findOneAndUpdate(
     { _id: contactId, owner },
     req.body
@@ -56,6 +75,7 @@ const updateById = async (req, res) => {
   if (!result) {
     throw HttpError(404, `Contact with id=${contactId} not found`);
   }
+
   res.json(result);
 };
 
@@ -63,9 +83,11 @@ const updateStatusContact = async (req, res) => {
   const { contactId } = req.params;
   const { favorite } = req.body;
   const { _id: owner } = req.user;
+
   if (!favorite) {
     return res.status(400, 'missing field favorite');
   }
+
   // const result = await Contact.findByIdAndUpdate(contactId, req.body);
   const result = await Contact.findOneAndUpdate(
     { _id: contactId, owner },
@@ -74,17 +96,20 @@ const updateStatusContact = async (req, res) => {
   if (!result) {
     throw HttpError(404, `Contact with id=${contactId} not found`);
   }
+
   res.json(result);
 };
 
 const deleleteById = async (req, res) => {
   const { contactId } = req.params;
   const { _id: owner } = req.user;
+
   // const result = await Contact.findByIdAndDelete(contactId);
   const result = await Contact.findOneAndDelete({ _id: contactId, owner });
   if (!result) {
     throw HttpError(404, `Contact with id=${contactId} not found`);
   }
+
   res.json({ message: 'Delete success' });
   // статус 204 не отправляет тело
   // res.status(204).json({ message: 'Delete success' });
